src/ReaderRFID/requirements.txt
paho-mqtt==1.2
RPi.GPIO==0.6.3
pi-rc522==1.1.0